requirements.txt
discord.py~=1.5.0
cryptography~=3.1.1
PyYAML~=5.3.1
SQLAlchemy~=1.3.20